devide codes of hubViewModel to some ViewModel
