+++ app/src/main/java/com/segnities007/seg/ui/components/top_bar/TopBar.kt
@@ -31,7 +31,7 @@ fun TopBar(
 ){
 
     CenterAlignedTopAppBar(
-            modifier = modifier.shadow(elevation = dimensionResource(R.dimen.padding_moderate)),
+            modifier = modifier.shadow(elevation = dimensionResource(R.dimen.padding_sn)),
             title = { Text(text = title, maxLines = 1,) },
             colors = TopAppBarDefaults.centerAlignedTopAppBarColors(
                 containerColor = MaterialTheme.colorScheme.primaryContainer,
