+++ app/src/main/res/values/dimens.xml
@@ -4,10 +4,12 @@
     <dimen name="padding_smallest">2dp</dimen>
     <dimen name="padding_smaller">4dp</dimen>
     <dimen name="padding_small">8dp</dimen>
-    <dimen name="padding_moderate">12dp</dimen>
+    <dimen name="padding_sn">12dp</dimen>
     <dimen name="padding_normal">16dp</dimen>
     <dimen name="padding_nl">24dp</dimen>
     <dimen name="padding_large">32dp</dimen>
+    <dimen name="padding_larger">64dp</dimen>
+    <dimen name="padding_largest">128dp</dimen>
     <dimen name="rounded_corner_shape">8dp</dimen>
     <dimen name="icon_small">48dp</dimen>
     <dimen name="icon_normal">64dp</dimen>
@@ -15,7 +17,7 @@
     <dimen name="icon_large">128dp</dimen>
     <dimen name="input_size">32dp</dimen>
     <dimen name="button_height_small_size">48dp</dimen>
-    <dimen name="button_height_normal_size">64dp</dimen>
+    <dimen name="button_height_normal_size">56dp</dimen>
     <dimen name="elevation_smallest">0.1dp</dimen>
     <dimen name="elevation_smaller">0.5dp</dimen>
     <dimen name="elevation_small">1dp</dimen>
